chore: move shared code to shared folder
