+++ src/node.js
@@ -1,17 +1,14 @@
-const AWS = require("aws-sdk");
-const { DynamoDB } = require("@aws-sdk/client-dynamodb");
-const { REGION } = require("./config");
+const { REGION } = require("./shared/config");
+const { getV2Response, getV3Response } = require("./shared/utils");
 
 (async () => {
   let response;
 
-  const v2Client = new AWS.DynamoDB({ region: REGION });
-  response = await v2Client.listTables().promise();
+  response = await getV2Response({ region: REGION });
   console.log("Data returned by v2:");
   console.log(JSON.stringify(response, null, 2));
 
-  const v3Client = new DynamoDB({ region: REGION });
-  response = await v3Client.listTables({});
+  response = await getV3Response({ region: REGION });
   console.log("\nData returned by v3:");
   console.log(JSON.stringify(response, null, 2));
 })();
